requirements.txt
altgraph==0.17.3
apturl==0.5.2
auto-py-to-exe==2.36.0
ayatana-settings==21.1.28
bcrypt==3.2.0
blinker==1.4
bottle==0.12.25
bottle-websocket==0.2.9
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.1
cajarename==21.11.24
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
deja-dup-caja==0.0.9
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
Eel==0.16.0
et-xmlfile==1.1.0
fasteners==0.14.1
Flask==2.3.3
folder-color-caja==0.0.86
folder-color-common==0.0.86
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gpg==1.16.0
greenlet==2.0.2
gspread==5.9.0
httplib2==0.20.2
idna==3.4
ImageHash==4.3.1
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
Magnus==1.0.3
Mako==1.1.3
MarkupSafe==2.0.1
mate-hud==22.4.4
mate-tweak==22.4.8
monotonic==1.6
more-itertools==8.10.0
mutagen==1.45.1
netifaces==0.11.0
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.8.0.74
openpyxl==3.1.2
pandas==2.0.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
psutil==5.9.0
ptyprocess==0.7.0
pulsemixer==1.5.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyGObject==3.42.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
PySimpleGUI==4.60.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-xlib==0.29
pytz==2023.3
PyWavelets==1.4.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
systemd-python==234
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vboxapi==1.0
wadllib==1.3.6
websockets==9.1
Werkzeug==2.3.7
whichcraft==0.6.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
yt-dlp==2022.4.8
zipp==1.0.0
zope.event==5.0
zope.interface==6.0
